basic layout

+++ app/models/product.rb
@@ -2,6 +2,6 @@ class Product
   include Mongoid::Document
   field :title, :type => String
   field :description, :type => String
-  field :price, :type => Decimal
+  field :price, :type => Integer
   field :image, :type => String
 end
